LP: #1821220 - Prevent base ESXi partitions from being modified.
The VMware ESXi image is a DD image which contains a base set of 8 partitions
which may not be modified. This ensures onces the VMFS6 storage layout a
user cannot delete, format, or mount any of those partitions. To remove
them a different layout must be applied first.

VMware ESXi also skips partition 4 in its layout. MAAS was not properly
skipping this partition, this has been fixed.

+++ src/maasserver/models/partition.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2016 Canonical Ltd.  This software is licensed under the
+# Copyright 2015-2019 Canonical Ltd.  This software is licensed under the
 # GNU Affero General Public License version 3 (see the file LICENSE).
 
 """Model for a partition in a partition table."""
@@ -198,6 +198,9 @@ class Partition(CleanSave, TimestampedModel):
 
     def get_partition_number(self):
         """Return the partition number in the table."""
+        # Avoid circular imports.
+        from maasserver.storage_layouts import VMFS6StorageLayout
+
         # Sort manually instead of with `order_by`, this will prevent django
         # from making a query if the partitions are already cached.
         partitions_in_table = self.partition_table.partitions.all()
@@ -213,16 +216,22 @@ class Partition(CleanSave, TimestampedModel):
             boot_disk = node.get_boot_disk()
             bios_boot_method = node.get_bios_boot_method()
             block_device = self.partition_table.block_device
-            if (arch == "ppc64el" and block_device.id == boot_disk.id):
+            vmfs_layout = VMFS6StorageLayout(self.get_node())
+            vmfs_bd = vmfs_layout.is_layout()
+            if vmfs_bd is not None:
+                # VMware ESXi is a DD image but MAAS allows partitions to
+                # be added to the end of the disk as well as resize the
+                # datastore partition. The EFI partition is already in the
+                # image so there is no reason to account for it.
+                if vmfs_bd.id == block_device.id and idx >= 3:
+                    # VMware ESXi skips the 4th partition.
+                    return idx + 2
+                else:
+                    return idx + 1
+            elif (arch == "ppc64el" and block_device.id == boot_disk.id):
                 return idx + 2
             elif arch == "amd64" and bios_boot_method != "uefi":
-                if node.osystem == 'esxi':
-                    # VMware ESXi is a DD image but MAAS allows partitions to
-                    # be added to the end of the disk as well as resize the
-                    # datastore partition. The EFI partition is already in the
-                    # image so there is no reason to account for it.
-                    return idx + 1
-                elif block_device.type == 'physical':
+                if block_device.type == 'physical':
                     # Delay the `type` check because it can cause a query. Only
                     # physical block devices get the bios_grub partition.
                     return idx + 2
@@ -309,6 +318,24 @@ class Partition(CleanSave, TimestampedModel):
                             "the MBR 2TiB maximum."],
                         })
 
+    def is_vmfs_partition(self):
+        # Avoid circular imports.
+        from maasserver.storage_layouts import VMFS6StorageLayout
+
+        vmfs_layout = VMFS6StorageLayout(self.get_node())
+        vmfs_bd = vmfs_layout.is_layout()
+        if vmfs_bd is None:
+            return False
+        if vmfs_bd.id != self.partition_table.block_device_id:
+            return False
+        if self.get_partition_number() >= len(vmfs_layout.base_partitions) + 2:
+            # A user may apply the VMFS6 layout and leave space at the end of
+            # the disk for additional VMFS datastores. Those partitions may be
+            # deleted, the base partitions may not as they are part of the DD.
+            # The + 2 is to account for partition 4 being skipped.
+            return False
+        return True
+
     def delete(self):
         """Delete the partition.
 
@@ -322,6 +349,10 @@ class Partition(CleanSave, TimestampedModel):
                 raise ValidationError(
                     "Cannot delete partition because its part of "
                     "a %s." % filesystem_group.get_nice_name())
+        if self.is_vmfs_partition():
+            raise ValidationError(
+                "VMware ESXi partitions may not be removed. To remove select "
+                "a different storage layout.")
         super(Partition, self).delete()
 
     def add_tag(self, tag):
